requirements/ci.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/ci.txt requirements/base.txt requirements/test-tools.in
#
amqp==2.6.0               # via -r requirements/base.txt, kombu
beautifulsoup4==4.9.1     # via webtest
billiard==3.6.3.0         # via -r requirements/base.txt, celery
celery==4.4.5             # via -r requirements/base.txt, flower
certifi==2020.4.5.1       # via -r requirements/base.txt, elastic-apm, requests, sentry-sdk
cffi==1.14.0              # via -r requirements/base.txt, cryptography
chardet==3.0.4            # via -r requirements/base.txt, requests
coverage==4.5.4           # via -r requirements/test-tools.in
cryptography==2.9.2       # via -r requirements/base.txt, django-auth-adfs
cssselect==1.1.0          # via pyquery
django-admin-index==1.4.0  # via -r requirements/base.txt
django-appconf==1.0.4     # via -r requirements/base.txt, django-axes, django-timeline-logger
django-auth-adfs-db==0.2.0  # via -r requirements/base.txt
django-auth-adfs==1.3.1   # via -r requirements/base.txt, django-auth-adfs-db
django-axes==5.3.2        # via -r requirements/base.txt
django-capture-on-commit-callbacks==1.1.0  # via -r requirements/test-tools.in
django-choices==1.7.1     # via -r requirements/base.txt, zgw-consumers
django-compat==1.0.15     # via -r requirements/base.txt, django-hijack, django-hijack-admin
django-extra-views==0.13.0  # via -r requirements/base.txt
django-filter==2.3.0      # via -r requirements/base.txt
django-fsm-admin==1.2.4   # via -r requirements/base.txt
django-fsm==2.7.0         # via -r requirements/base.txt, django-fsm-admin
django-hijack-admin==2.1.10  # via -r requirements/base.txt
django-hijack==2.1.10     # via -r requirements/base.txt, django-hijack-admin
django-ipware==2.1.0      # via -r requirements/base.txt, django-axes
django-jenkins==0.110.0   # via -r requirements/test-tools.in
django-ordered-model==3.4.1  # via -r requirements/base.txt, django-admin-index
django-privates==1.2.1    # via -r requirements/base.txt
django-redis==4.12.1      # via -r requirements/base.txt
django-relativedelta==1.1.1  # via -r requirements/base.txt, zgw-consumers
django-rosetta==0.9.4     # via -r requirements/base.txt
django-sendfile2==0.6.0   # via -r requirements/base.txt, django-privates
django-sniplates==0.7.0   # via -r requirements/base.txt
django-solo==1.1.3        # via -r requirements/base.txt, django-auth-adfs-db, zgw-consumers
django-timeline-logger==1.1.2  # via -r requirements/base.txt
django-webtest==1.9.7     # via -r requirements/test-tools.in
django==2.2.12            # via -r requirements/base.txt, django-appconf, django-auth-adfs, django-auth-adfs-db, django-axes, django-capture-on-commit-callbacks, django-choices, django-extra-views, django-filter, django-fsm-admin, django-jenkins, django-privates, django-redis, django-relativedelta, django-rosetta, django-sendfile2, django-sniplates, django-timeline-logger, zgw-consumers
elastic-apm==5.6.0        # via -r requirements/base.txt
factory-boy==2.12.0       # via -r requirements/test-tools.in
faker==4.1.0              # via factory-boy
flower==0.9.4             # via -r requirements/base.txt
freezegun==0.3.15         # via -r requirements/test-tools.in
future==0.18.2            # via -r requirements/base.txt, celery
gemma-zds-client==0.13.1  # via -r requirements/base.txt, zgw-consumers
humanize==0.5.1           # via -r requirements/base.txt, flower
idna==2.9                 # via -r requirements/base.txt, requests
kombu==4.6.10             # via -r requirements/base.txt, celery
lxml==4.6.2               # via pyquery
pillow==7.1.2             # via -r requirements/base.txt
polib==1.1.0              # via -r requirements/base.txt, django-rosetta
psycopg2==2.8.5           # via -r requirements/base.txt
pycparser==2.20           # via -r requirements/base.txt, cffi
pyjwt==1.7.1              # via -r requirements/base.txt, django-auth-adfs, gemma-zds-client
pyquery==1.4.1            # via -r requirements/test-tools.in
python-dateutil==2.8.1    # via -r requirements/base.txt, django-relativedelta, faker, freezegun
python-decouple==3.3      # via -r requirements/base.txt
python-dotenv==0.13.0     # via -r requirements/base.txt
pytz==2020.1              # via -r requirements/base.txt, celery, django, flower
pyyaml==5.3.1             # via -r requirements/base.txt, gemma-zds-client
redis==3.5.3              # via -r requirements/base.txt, django-redis
requests-mock==1.8.0      # via -r requirements/test-tools.in
requests==2.23.0          # via -r requirements/base.txt, django-auth-adfs, django-rosetta, gemma-zds-client, requests-mock
sentry-sdk==0.14.4        # via -r requirements/base.txt
six==1.15.0               # via -r requirements/base.txt, cryptography, django-choices, django-compat, django-extra-views, django-rosetta, freezegun, python-dateutil, requests-mock, webtest
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.3.1           # via -r requirements/base.txt, django
tblib==1.6.0              # via -r requirements/test-tools.in
text-unidecode==1.3       # via faker
tornado==6.0.4            # via -r requirements/base.txt, flower
urllib3==1.25.9           # via -r requirements/base.txt, elastic-apm, requests, sentry-sdk
uwsgi==2.0.18             # via -r requirements/base.txt
vine==1.3.0               # via -r requirements/base.txt, amqp, celery
waitress==1.4.4           # via webtest
webob==1.8.6              # via webtest
webtest==2.0.35           # via django-webtest
zgw-consumers==0.10.2     # via -r requirements/base.txt
